Net7: There is no ssize_t on Windows.

+++ src/libs/3rdparty/net7ssh/src/ne7ssh_transport.cpp
@@ -194,7 +194,7 @@ bool ne7ssh_transport::wait (SOCKET socket, int rw, int timeout)
 
 bool ne7ssh_transport::send (Botan::SecureVector<Botan::byte>& buffer)
 {
-  ssize_t byteCount;
+  long byteCount;
   uint32 sent = 0;
 
   if (buffer.size() > MAX_PACKET_LEN)
